+++ NEWS.md
@@ -1,3 +1,10 @@
+# ver. 1.0.0.9002
+* Continuous development version.
+* New features updated: 
+    + Now user can flip the input value of meta study for calculating exp(Value) and confidence interval; e.g. when meta beta is < 0, let's say -0.5, setting **flipMeta** = TRUE will let downstream calculation use -(meta beta), which is -(-0.5) = 0.5, or vice versa.
+    + Now forest plot reference line on the non-exponentiated scale is more sensible.
+
+
 # ver. 1.0.0.9001
 * Continuous development version.
 * New features updated: